+++ packages/clipper/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.72](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.71...@standardnotes/clipper@1.1.72) (2023-07-05)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.71](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.70...@standardnotes/clipper@1.1.71) (2023-07-05)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.70](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.69...@standardnotes/clipper@1.1.70) (2023-07-05)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.69](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.68...@standardnotes/clipper@1.1.69) (2023-07-04)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.68](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.67...@standardnotes/clipper@1.1.68) (2023-07-04)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
 ## [1.1.67](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.66...@standardnotes/clipper@1.1.67) (2023-07-04)
 
 **Note:** Version bump only for package @standardnotes/clipper