+++ packages/files/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.16.2](https://github.com/standardnotes/app/compare/@standardnotes/files@1.16.1...@standardnotes/files@1.16.2) (2023-07-05)
+
+**Note:** Version bump only for package @standardnotes/files
+
+## [1.16.1](https://github.com/standardnotes/app/compare/@standardnotes/files@1.16.0...@standardnotes/files@1.16.1) (2023-07-04)
+
+**Note:** Version bump only for package @standardnotes/files
+
 # [1.16.0](https://github.com/standardnotes/app/compare/@standardnotes/files@1.15.15...@standardnotes/files@1.16.0) (2023-07-03)
 
 ### Features